feat (Crear y Modificar curso): Se agregan las funcionalidades para crear y eliminar curso, pendiente el modificar y algunas partes como etiquetas y el singleton usuario

+++ Vistas/CalificacionCursoPagina.xaml.cs
@@ -113,7 +113,7 @@ namespace UVemyCliente.Vistas
             grdBackground.IsEnabled = false;
 
             int calificacionNueva = Int32.Parse(txtBlockCalificacion.Text);
-            UsuarioCursoDTO usuario = new UsuarioCursoDTO { IdCurso = _curso.IdCurso, IdUsuario = 3, Calificacion = calificacionNueva };
+            UsuarioCursoDTO usuario = new UsuarioCursoDTO { IdCurso = (int)_curso.IdCurso, IdUsuario = 3, Calificacion = calificacionNueva };
             var json = JsonSerializer.Serialize(usuario);
             var content = new StringContent(json, Encoding.UTF8, "application/json");
